\name{calfs}
\alias{calfs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calfs}
\format{A raster image array with red, green, and blue channels.}
\description{
Base image is original by the author.
}
\source{\url{https://flic.kr/p/GUFEis}}
\author{
Christopher Steven Marcum
}
\keyword{ ~data}
