\name{plot.halftone}
\alias{plot.halftone}
\alias{points.halftone}
\title{S3 plotting methods for halftone class objects.}
\description{Plots points on a grid passed from a halftone class object.}
\usage{
\method{plot}{halftone}(x,\dots)
\method{points}{halftone}(x,\dots)
}
\arguments{
\item{x}{A halftone object.}
\item{\dots}{Additional arguments passed to plotting routines except for cex, which is inferred from x.}
}
\keyword{methods}
